// web/src/shared/api/generated/model/avatarCitationResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { LessonContentSection } from './lessonContentSection';

export interface AvatarCitationResult {
  reference: string;
  section: LessonContentSection;
  /** @nullable */
  sectionTitle: string | null;
  lessonId: string;
  /** @nullable */
  questionId: string | null;
}