// web/src/shared/api/generated/model/lessonContentSection.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export type LessonContentSection = typeof LessonContentSection[keyof typeof LessonContentSection];


export const LessonContentSection = {
  Explanation: 'Explanation',
  Objectives: 'Objectives',
  Summary: 'Summary',
  QuestionExplanation: 'QuestionExplanation',
} as const;